+++ src/main.rs
@@ -1,16 +1,8 @@
 // Copyright 2018-2020 Sebastian Wiesner <sebastian@swsnr.de>
 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-
-// 	http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #![deny(warnings, clippy::all)]
 
@@ -146,8 +138,12 @@ markdown documents for viewing in text terminals:
 • In some terminals: Inline images and inline links
 • In iTerm2: Jump marks for headings
 
-Copyright (C) 2018–2020 Sebastian Wiesner and contributors
-Licensed under the Apache License, Version 2.0
+Copyright (C) Sebastian Wiesner and contributors
+
+This program is subject to the terms of the Mozilla Public License, 
+v. 2.0. If a copy of the MPL was not distributed with this file,
+You can obtain one at http://mozilla.org/MPL/2.0/.
+
 Report issues to <https://github.com/lunaryorn/mdcat>.",
         )
         .arg(